+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Version number is based on [Semver](https://semver.org/).
 Note that major version zero (0.y.z) is for initial development and anything may change at any time.
 
+## [Unreleased]
+
+### Added
+- Show mean and median grade in plot
+- Add selection to show raw percentage in the plot
+
+### Fixed
+- Make final notes easier to read
+
 ## [0.3.3] - 2021-11-12
 
 ### Added