+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Note that major version zero (0.y.z) is for initial development and anything may
 - Make saved messages stand out and simplify final info note.
 - Make hover highlighting more intuitive and better explained.
 - Filter out unpublished assignment and those missing a max score.
+- Put assignment titles on top of plots instead of to the left.
 
 ### Fixed
 - Make rounding work with dfs containing None instead of NaN