+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Version number is based on [Semver](https://semver.org/).
 Note that major version zero (0.y.z) is for initial development and anything may change at any time.
 
+## [0.5.1] 2022-10-01
+
+### Changed
+- Remove placeholder commands that have been implemented as part of viz already.
+
 ## [0.5.0] 2022-09-29
 
 ### Added